Reconstruct remaining actors functions

+++ source/ai/actor_placement.h
@@ -9,7 +9,7 @@ struct actor_starting_location;
 long actor_place(
 	long actor_variant_definition_index,
 	long encounter_index,
-	short squad_index,
+	long squad_index,
 	struct actor_starting_location *starting_location,
 	boolean upgrade_major,
 	short initial_variant);
